Added files from private datanavy search repo
All the files in this commit are from private datanavy search repository. For some reasons, original
repository cannot be made public. Datanavy search was once hosted at domain www.datanavy.site on
www.pythonanywhere.com. This repo is meant to be public version of the private repository which
was created back in 2021.

The files were tested, they may not work as some files were removed. Spacy takes long to load thats why I
havent tested them but in comming commits they will be tested. Code in the files is not documented and there are
no comments or doctrings to help understand the code. If there are errors I will fix them but I wont maintain it
ever. Even the private repo was not maintained from 2021 because the code is not reusable and its hard to fix even
small bug as it requires changing other parts.

The files include python, html and css files which is expected as datanavy was running as webpage.
Code is both complex at client and server which make it hard to maintain. I was lazy to change the code when
it was small but realised that it was hard to do it when it was large. Maintaining the project is not worth
but the code can be refactored to be ussed in seperate repository.

I was still learning Python when creating the repository and could wait to code with. Little did I know I had
to face SQL, client-server technologies, AngularJS and others. The only thing I cared was creating application
and never cared about readability and maintainance. I thought that I was going to finish the project in few weeks
but took months up until I quited.

It took lot efford and time to create that code but know I realised that lot could have been avoided easily with
code reuse and planning. I realise that, starting to code without understanding the problem statement and scode
would just cause problems. I wanted to create application to search but I didnt know that it was search engine.
Most of issues was not with languages(css, html, etc) but with not understanding the problem. Most of the files
were rewritten from scratch multiple times because of changes in goals.

Pynavy and Cnavy search are under development that are based on this code but with different languages. Pynavy
just cleans this code into maintainable and optimised code. Cnavy uses C\C++ but is little different from
datanavy search.

Focus is now on Pynavy and Cnavy search an dthis repo will act as reference for Pynavy search.