Fixed small errors
The application was running but on Edge it wasnt but it can be done with babel transpiler. Sometimes webpage
wasnt fully loaded while other times loaded.This may be related to network issues on my side but could also
be the application issue. Problem is that, some html files are loaded dynamically by AngularJS and if html with
that AngularJS fails to load or disturbe dit causes the whole webpage to break.

There are just a lot of issues on client side. Fixing is just a waste of time as new issues will just pop in
until and until. It works but works by chance, sometimes the webpage may not be fully loaded which may require
reloaded and it could also fail to fully load.

Most fixes was to create users/ directory automatically and install Spacy small english mode and wordnet
automatically without having to execute Python commands seperatively.

The code in this repo is a hell.